refactored for manifests
