--- .idea/encodings.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="Encoding">
-    <file url="file://$PROJECT_DIR$/tmp/server/logs/valheim_server.log" charset="windows-1252" />
-  </component>
-</project>
