--- .idea/externalDependencies.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="ExternalDependencies">
-    <plugin id="org.rust.lang" />
-    <plugin id="org.toml.lang" />
-  </component>
-</project>
