--- .idea/remote-targets.xml
@@ -1,22 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="RemoteTargetsManager">
-    <option name="projectDefaultTargetUuid" value="af4ba5d3-9923-40ed-a296-b27f43a88e09" />
-    <targets>
-      <target name="WSL - Ubuntu-22.04" type="wsl" uuid="af4ba5d3-9923-40ed-a296-b27f43a88e09">
-        <config>
-          <option name="distributionMsId" value="Ubuntu" />
-          <option name="projectRootOnTarget" value="/mnt/c/Users/micha/AppData/Local/JetBrains/Toolbox/apps/IDEA-U/ch-0/223.8214.52/jbr/bin&#10;/valheim-docker" />
-        </config>
-        <ContributedStateBase type="RsLanguageRuntime">
-          <config>
-            <option name="cargoPath" value="/home/mbruno/.cargo/bin/cargo" />
-            <option name="cargoVersion" value="1.66.1 (ad779e08b 2023-01-10)" />
-            <option name="rustcPath" value="/home/mbruno/.cargo/bin/rustc" />
-            <option name="rustcVersion" value="1.66.1 (90743e729 2023-01-10)" />
-          </config>
-        </ContributedStateBase>
-      </target>
-    </targets>
-  </component>
-</project>
