+++ src/huginn/Cargo.toml
@@ -22,7 +22,7 @@ path = "main.rs"
 cc = "1"
 
 [dependencies]
-log = "0.4.25"
+log = "0.4.26"
 tokio = { version = "1", features = ["full"] }
 warp = { version = "0.3.7" }
