+++ src/odin/Cargo.toml
@@ -31,14 +31,14 @@ cc = "1"
 [dependencies]
 handlebars = "6.3.1"
 dotenv = "0.15"
-log = "0.4.25"
-clap = { version = "4.5.28", features = [ "derive", "env" ] }
+log = "0.4.26"
+clap = { version = "4.5.30", features = [ "derive", "env" ] }
 which = "*"
-serde = { version = "1.0.217", features = ["derive"], default-features = false  }
+serde = { version = "1.0.218", features = ["derive"], default-features = false  }
 sysinfo = { version = "0.33.1" }
-serde_json = "1.0.138"
+serde_json = "1.0.139"
 daemonize = "0.5.0"
-tar = "0.4.43"
+tar = "0.4.44"
 flate2 = "1.0.35"
 inflections = "1.1.1"
 md5 = "0.7"
@@ -54,7 +54,8 @@ tokio = { version = "1", features = ["full"] }
 notify = "8.0.0"
 json-patch = "*"
 cached = "0"
-anyhow = "1.0.95"
+anyhow = "1.0.96"
+thiserror = "2.0.11"
 
 [dev-dependencies]
 once_cell = "1.20.3"
