+++ src/odin/server/mod.rs
@@ -5,5 +5,6 @@ mod startup;
 mod status;
 mod update;
 mod utils;
+
 // Reexport all public functions
 pub use crate::server::{install::*, shutdown::*, startup::*, status::*, update::*, utils::*};
